Fix the style issues 1. spacing format 2. merge *_params.h into dataset_test_base 3. function argument comment format
