+++ tensorflow/core/framework/function_testlib.h
@@ -143,26 +143,26 @@ FunctionDef XYXLessThanOrEqualToN(int64 N);
 // x: T -> bool
 FunctionDef RandomUniformLess();
 
-// start:int64, stop:int64, step:int64 -> y: RangeDatasetOp::Dataset
+// start:int64, stop:int64, step:int64 -> y:RangeDatasetOp::Dataset
 FunctionDef MakeRangeDataset();
 
-// input_dataset: variant, batch_size: int64, drop_remainder: bool
-// -> y: BatchDatasetV2::Dataset
+// input_dataset:variant, batch_size:int64, drop_remainder:bool
+// -> y:BatchDatasetV2::Dataset
 FunctionDef MakeBatchDataset();
 
-// input_dataset: variant, other_arguments: Targuments, f: func,
-// Targuments: list(type), output_types: list(type), output_shapes: list(shape),
-// use_inter_op_parallelism: bool, preserve_cardinality: bool
-// -> y: MapDatasetOp::Dataset
+// input_dataset:variant, other_arguments:Targuments, f:func,
+// Targuments:list(type), output_types:list(type), output_shapes:list(shape),
+// use_inter_op_parallelism:bool, preserve_cardinality:bool
+// -> y:MapDatasetOp::Dataset
 FunctionDef MakeMapDataset(bool has_other_args);
 
-// input_dataset:variant, count:int64 -> y: TakeDataset::Dataset
+// input_dataset:variant, count:int64 -> y:TakeDataset::Dataset
 FunctionDef MakeTakeDataset();
 
-// x:T -> y: TensorSliceDatasetOp::Dataset
+// x:T -> y:TensorSliceDatasetOp::Dataset
 FunctionDef MakeTensorSliceDataset();
 
-// x:T -> y: T, idx: out_idx
+// x:T -> y:T, idx:out_idx
 FunctionDef Unique();
 
 void FunctionTestSchedClosure(std::function<void()> fn);
