+++ tensorflow/core/kernels/data/dataset_test_base.cc
@@ -696,6 +696,7 @@ Status DatasetOpsTestBaseV2::Initialize(DatasetParams& dataset_params) {
   TF_RETURN_IF_ERROR(CreateIteratorContext(dataset_ctx_.get(), &iterator_ctx_));
   TF_RETURN_IF_ERROR(dataset_->MakeIterator(
       iterator_ctx_.get(), dataset_params.iterator_prefix(), &iterator_));
+  return Status::OK();
 }
 
 Status DatasetOpsTestBaseV2::MakeDatasetOpKernel(
@@ -729,12 +730,6 @@ Status DatasetOpsTestBaseV2::MakeDatasetTensor(DatasetParams* dataset_params,
   std::vector<Tensor> input_tensors;
   for (auto& tensor_value : inputs) {
     input_tensors.emplace_back(*tensor_value.tensor);
-  }
-
-  bool has_other_args = false;
-  if (dataset_params->type() == DatasetParamsType::Map) {
-    auto map_dataset_params = dynamic_cast<MapDatasetParams*>(dataset_params);
-    has_other_args = map_dataset_params->num_of_other_arguments() > 0;
   }
 
   GraphConstructorOptions graph_opts;
@@ -771,6 +766,142 @@ Status DatasetOpsTestBaseV2::MakeDatasetTensorFunc(
                                    ToString(dataset_params.type()));
   }
   return Status::OK();
+}
+
+string ToString(DatasetParamsType type) {
+  switch (type) {
+    case DatasetParamsType::Range:
+      return "Range";
+    case DatasetParamsType::Batch:
+      return "Batch";
+    case DatasetParamsType::Map:
+      return "Map";
+  }
+}
+
+DatasetParams::DatasetParams(DataTypeVector output_dtypes,
+                             std::vector<PartialTensorShape> output_shapes,
+                             string node_name, DatasetParamsType type)
+    : output_dtypes_(std::move(output_dtypes)),
+      output_shapes_(std::move(output_shapes)),
+      node_name_(std::move(node_name)),
+      type_(type) {}
+
+bool DatasetParams::IsDatasetTensor(const Tensor& tensor) {
+  return tensor.dtype() == DT_VARIANT &&
+         TensorShapeUtils::IsScalar(tensor.shape());
+}
+
+RangeDatasetParams::RangeDatasetParams(
+    int64 start, int64 stop, int64 step, DataTypeVector output_dtypes,
+    std::vector<PartialTensorShape> output_shapes, string node_name)
+    : DatasetParams(std::move(output_dtypes), std::move(output_shapes),
+                    std::move(node_name), DatasetParamsType::Range),
+      start_(CreateTensor<int64>(TensorShape({}), {start})),
+      stop_(CreateTensor<int64>(TensorShape({}), {stop})),
+      step_(CreateTensor<int64>(TensorShape({}), {step})) {}
+
+RangeDatasetParams::RangeDatasetParams(int64 start, int64 stop, int64 step)
+    : DatasetParams({DT_INT64}, {PartialTensorShape({})}, "range_dataset",
+                    DatasetParamsType::Range),
+      start_(CreateTensor<int64>(TensorShape({}), {start})),
+      stop_(CreateTensor<int64>(TensorShape({}), {stop})),
+      step_(CreateTensor<int64>(TensorShape({}), {step})) {}
+
+Status RangeDatasetParams::MakeInputs(
+    gtl::InlinedVector<TensorValue, 4>* inputs) {
+  *inputs = {TensorValue(&start_), TensorValue(&stop_), TensorValue(&step_)};
+  return Status::OK();
+}
+
+Status RangeDatasetParams::MakeInputPlaceholder(
+    std::vector<string>* input_placeholder) const {
+  *input_placeholder = {RangeDatasetOp::kStart, RangeDatasetOp::kStop,
+                        RangeDatasetOp::kStep};
+  return Status::OK();
+}
+
+Status RangeDatasetParams::MakeAttributes(AttributeVector* attr_vector) const {
+  *attr_vector = {{RangeDatasetOp::kOutputTypes, output_dtypes_},
+                  {RangeDatasetOp::kOutputShapes, output_shapes_}};
+  return Status::OK();
+}
+
+Status BatchDatasetParams::MakeInputs(
+    gtl::InlinedVector<TensorValue, 4>* inputs) {
+  inputs->reserve(input_dataset_params_group_.size());
+  for (auto& pair : input_dataset_params_group_) {
+    if (!IsDatasetTensor(pair.second)) {
+      inputs->clear();
+      return errors::Internal(
+          "The input dataset is not populated as the dataset tensor yet.");
+    } else {
+      inputs->emplace_back(TensorValue(&pair.second));
+    }
+  }
+  inputs->emplace_back(TensorValue(&batch_size_));
+  inputs->emplace_back(TensorValue(&drop_remainder_));
+  return Status::OK();
+}
+
+Status BatchDatasetParams::MakeInputPlaceholder(
+    std::vector<string>* input_placeholder) const {
+  *input_placeholder = {BatchDatasetOp::kInputDataset,
+                        BatchDatasetOp::kBatchSize,
+                        BatchDatasetOp::kDropRemainder};
+  return Status::OK();
+}
+
+Status BatchDatasetParams::MakeAttributes(AttributeVector* attr_vector) const {
+  *attr_vector = {{BatchDatasetOp::kParallelCopy, parallel_copy_},
+                  {BatchDatasetOp::kOutputTypes, output_dtypes_},
+                  {BatchDatasetOp::kOutputShapes, output_shapes_}};
+  return Status::OK();
+}
+
+int BatchDatasetParams::op_version() const { return op_version_; }
+
+Status MapDatasetParams::MakeInputs(
+    gtl::InlinedVector<TensorValue, 4>* inputs) {
+  inputs->reserve(input_dataset_params_group_.size());
+  for (auto& pair : input_dataset_params_group_) {
+    if (!IsDatasetTensor(pair.second)) {
+      inputs->clear();
+      return errors::Internal(
+          "The input dataset is not populated as the dataset tensor yet.");
+    } else {
+      inputs->emplace_back(TensorValue(&pair.second));
+    }
+  }
+  for (auto& argument : other_arguments_) {
+    inputs->emplace_back(TensorValue(&argument));
+  }
+  return Status::OK();
+}
+
+Status MapDatasetParams::MakeInputPlaceholder(
+    std::vector<string>* input_placeholder) const {
+  input_placeholder->emplace_back(MapDatasetOp::kInputDataset);
+  for (int i = 0; i < other_arguments_.size(); ++i) {
+    input_placeholder->emplace_back(
+        absl::StrCat(MapDatasetOp::kOtherArguments, "_", i));
+  }
+  return Status::OK();
+}
+
+Status MapDatasetParams::MakeAttributes(AttributeVector* attr_vector) const {
+  *attr_vector = {
+      {MapDatasetOp::kFunc, func_},
+      {MapDatasetOp::kTarguments, type_arguments_},
+      {MapDatasetOp::kOutputShapes, output_shapes_},
+      {MapDatasetOp::kOutputTypes, output_dtypes_},
+      {MapDatasetOp::kUseInterOpParallelism, use_inter_op_parallelism_},
+      {MapDatasetOp::kPreserveCardinality, preserve_cardinality_}};
+  return Status::OK();
+}
+
+std::vector<FunctionDef> MapDatasetParams::func_lib() const {
+  return func_lib_;
 }
 
 }  // namespace data
