+++ tensorflow/core/kernels/data/range_dataset_op_test.cc
@@ -15,7 +15,6 @@ limitations under the License.
 #include "tensorflow/core/kernels/data/range_dataset_op.h"
 
 #include "tensorflow/core/kernels/data/dataset_test_base.h"
-#include "tensorflow/core/kernels/data/range_dataset_op_test.h"
 
 namespace tensorflow {
 namespace data {
